shared/requirements.txt
# Shared Python Requirements
# Dependencies used across all agents

# Core libraries
pyyaml>=6.0.1
requests>=2.31.0
psutil>=5.9.5

# Configuration and validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Optional dependencies for advanced features
# Uncomment as needed

# GPU monitoring
# pynvml>=11.5.0

# Distributed computing
# ray>=2.8.0
# dask[complete]>=2023.12.0

# Monitoring and metrics
# prometheus-client>=0.19.0

# API framework (if building REST APIs)
# fastapi>=0.105.0
# uvicorn>=0.25.0
